created manifest store ABC, move previous manifest store to API manifest store
