+++ tests/test_z_integration.py
@@ -8,7 +8,7 @@ from schematic_db.rdb.synapse_database import SynapseDatabase
 from schematic_db.rdb_builder.rdb_builder import RDBBuilder
 from schematic_db.schema.schema import Schema
 from schematic_db.rdb_updater.rdb_updater import RDBUpdater
-from schematic_db.manifest_store.manifest_store import ManifestStore
+from schematic_db.manifest_store.api_manifest_store import APIManifestStore
 from schematic_db.query_store.query_store import QueryStore
 from schematic_db.rdb_queryer.rdb_queryer import RDBQueryer
 
@@ -45,7 +45,7 @@ def fixture_rdb_builder_synapse(
 
 @pytest.fixture(scope="class", name="rdb_updater_mysql")
 def fixture_rdb_updater_mysql(
-    mysql_database: MySQLDatabase, manifest_store: ManifestStore
+    mysql_database: MySQLDatabase, manifest_store: APIManifestStore
 ) -> Generator:
     """Yields a RDBUpdater with a mysql database and test schema"""
     obj = RDBUpdater(rdb=mysql_database, manifest_store=manifest_store)
@@ -55,7 +55,7 @@ def fixture_rdb_updater_mysql(
 
 @pytest.fixture(scope="class", name="rdb_updater_postgres")
 def fixture_rdb_updater_postgres(
-    postgres_database: PostgresDatabase, manifest_store: ManifestStore
+    postgres_database: PostgresDatabase, manifest_store: APIManifestStore
 ) -> Generator:
     """Yields a RDBUpdater with a mysql database and test schema"""
     obj = RDBUpdater(rdb=postgres_database, manifest_store=manifest_store)
@@ -65,7 +65,7 @@ def fixture_rdb_updater_postgres(
 
 @pytest.fixture(scope="class", name="rdb_updater_synapse")
 def fixture_rdb_updater_synapse(
-    synapse_database: SynapseDatabase, manifest_store: ManifestStore
+    synapse_database: SynapseDatabase, manifest_store: APIManifestStore
 ) -> Generator:
     """Yields a RDBUpdater with a synapse database and test schema"""
     obj = RDBUpdater(rdb=synapse_database, manifest_store=manifest_store)
